Migrate classification code into new controllers
Updating actions, paths and forms to keep the feature working.

Using a POST create action to save the new classification as we record
every time the state instead of just updating the previous state.
