+++ app/controllers/alaveteli_pro/info_requests_controller.rb
@@ -55,17 +55,6 @@ class AlaveteliPro::InfoRequestsController < AlaveteliPro::BaseController
     end
   end
 
-  def update
-    @info_request = InfoRequest.find(params[:id])
-    authorize! :update_request_state, @info_request
-    new_status = info_request_params[:described_state]
-    @info_request.set_described_state(new_status, current_user)
-    flash[:notice] = _('Your request has been updated!')
-    redirect_to show_alaveteli_pro_request_path(
-      url_title: @info_request.url_title
-    )
-  end
-
   private
 
   def show_errors
@@ -134,10 +123,6 @@ class AlaveteliPro::InfoRequestsController < AlaveteliPro::BaseController
     params.
       require(:alaveteli_pro_request_filter).
         permit(:filter, :order, :search)
-  end
-
-  def info_request_params
-    params.require(:info_request).permit(:described_state)
   end
 
   def check_public_body_is_requestable
