+++ spec/controllers/alaveteli_pro/info_requests_controller_spec.rb
@@ -100,22 +100,4 @@ describe AlaveteliPro::InfoRequestsController do
       end
     end
   end
-
-  describe '#update' do
-    let(:pro_user) { FactoryBot.create(:pro_user) }
-    let(:other_pro_user) { FactoryBot.create(:pro_user) }
-    let(:info_request) { FactoryBot.create(:info_request, user: pro_user) }
-
-    context 'when the user is not allowed to update the request' do
-      it 'raises a CanCan::AccessDenied error' do
-        session[:user_id] = other_pro_user.id
-        expect do
-          put :update, params: {
-            id: info_request.id,
-            info_request: { described_state: 'successful' }
-          }
-        end.to raise_error(CanCan::AccessDenied)
-      end
-    end
-  end
 end
